fix status

+++ controllers/film.controllers.js
@@ -1,27 +1,41 @@
 const {Films,Actor,FilmType, sequelize} = require("../models")
-
+const fs = require('fs');
 const createFilm = async (req,res)=>{
-    const data = req.body
+    const data = req.body  
     try {
         const newFilm = await Films.create(data);
         res.status(201).send(newFilm)
     } catch (error) {
-        res.status(505).send(error)
+        res.status(400).send(error)
     }
 }
 const deleteFilm = async (req,res)=>{
     const id = req.params.id;
     const ID = parseInt(id)
     try {
-        await Films.destroy({
-           where: {
-             id:ID
-           }
-         });
-         res.status(200).send("Delete successful !")
+        const result = await Films.findOne({
+            where: {
+              id:ID
+            }
+          });
+
+        if(result){
+            await Films.destroy({
+                where: {
+                  id:ID
+                }
+              });
+             fs.unlinkSync(result.img);
+             fs.unlinkSync(result.src);
+             fs.unlinkSync(result.trailer);
+             res.status(200).send("Delete successful !")
+        }else{
+            res.status(404).send("Not found !")
+        }
+        
        
    } catch (error) {
-       res.status(500).send(err)
+       res.status(400).send(err)
    }
 }
 const updateFilm = async (req,res)=>{
@@ -46,7 +60,7 @@ const updateFilm = async (req,res)=>{
         }
 
     } catch (error) {
-        res.status(505).send(error)
+        res.status(400).send(error)
     }
 }
 const getFilmUser = async (req,res)=>{
@@ -67,7 +81,7 @@ const getFilmUser = async (req,res)=>{
         res.status(200).send(result);
 
     } catch (error) {
-        res.status(505).send(error)
+        res.status(400).send(error)
     }
 }
 const getDetailFilm = async (req,res)=>{
@@ -100,6 +114,7 @@ const getDetailFilm = async (req,res)=>{
                 src:film.src,
                 status:film.status,
                 img:film.img,
+                trailer:film.trailer,
         }
         if(actor){
             film = {
@@ -117,10 +132,29 @@ const getDetailFilm = async (req,res)=>{
         
 
        }else{
-        res.status(400).send("Not found")
+        res.status(404).send("Not found")
        }
     } catch (error) {
-        res.status(500).send(error)
+        res.status(400).send(error)
+    }
+}
+
+const getFilmWatching = async (req,res)=>{
+    const id = req.params.id;
+    const numberId = parseInt(id)
+    try {
+        let film = await Films.findOne({
+            where:{
+                id:numberId
+            }
+        })
+        if(film){
+            res.status(200).send(film)
+        }else{
+            res.status(404).send("not found !")
+        }
+    } catch (error) {
+        res.status(400).send(error)
     }
 }
 
@@ -128,9 +162,13 @@ const getDetailFilm = async (req,res)=>{
 const getFilmAdmin = async (req,res)=>{
     try {
         const result = await Films.findAll();
-        res.status(200).send(result);
+        if(result){
+            res.status(200).send(result);
+        }else{
+            res.status(404).send("Not found !");
+        }
     } catch (error) {
-        res.status(500).send(error);
+        res.status(400).send(error);
     }
 }
 
@@ -144,10 +182,35 @@ const searchFilm = async (req,res)=>{
         if(result){
             res.status(200).send(result[0])
         }else{
-            res.status(400).send("Not found !")
+            res.status(404).send("Not found !")
         }
     } catch (error) {
-        res.status(500).send(error)
+        res.status(400).send(error)
+    }
+}
+
+const uploadFilm = async (req,res)=>{
+    const img = req.files.img[0].path;
+    const trailer = req.files.trailer[0].path;
+    const src = req.files.src[0].path;
+
+    const id = parseInt(req.params.id)
+    try {
+        const result = await Films.findOne({
+            where: {
+                id
+            }
+        })
+
+        result.src = src;
+        result.img = img;
+        result.trailer = trailer;
+
+        result.save()
+
+        res.status(201).send(result)
+    } catch (error) {
+        res.status(400).send(error)
     }
 }
 
@@ -158,5 +221,7 @@ module.exports = {
     getFilmAdmin,
     getFilmUser,
     getDetailFilm,
-    searchFilm
+    searchFilm,
+    uploadFilm,
+    getFilmWatching
 }
